backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0

# Audio Processing
pyaudio==0.2.11
numpy==1.24.3
scipy==1.11.4
librosa==0.10.1

# AI Services
openai==1.3.7
anthropic==0.7.7
google-cloud-translate==3.12.1
google-cloud-speech==2.21.0

# Local AI Models
torch==2.1.1
transformers==4.35.2
whisper==1.1.10
sentencepiece==0.1.99

# Translation
deep-translator==1.11.4
googletrans==4.0.0rc1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
aiofiles==23.2.1

# Logging & Monitoring
loguru==0.7.2
prometheus-client==0.19.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# Performance
redis==5.0.1
aioredis==2.0.1
asyncio-throttle==1.0.2
